guest : Removed ARM architecture dependencies in guest context
We have to integrate both files(guest.c, context.c). and create a hal
directory for arm32ve. After all, we plan to move there, regarding
the part of architecture-specific code.
