+++ hypervisor/interrupt.c
@@ -1,7 +1,7 @@
 #include "hvmm_types.h"
 #include "gic.h"
 #include "armv7_p15.h"
-#include "context.h"
+#include "guest.h"
 #include "hvmm_trace.h"
 #include "vgic.h"
 #include "virq.h"
@@ -34,5 +34,9 @@ hvmm_status_t hvmm_interrupt_init(void)
         ret = vgic_enable(1);
     if (ret == HVMM_STATUS_SUCCESS)
         ret = virq_init();
+
+    /* Initialize PIRQ to VIRQ mapping */
+    virqmap_init();
+
     return ret;
 }
