+++ hypervisor/vdev/vdev_hvc_ping.c
@@ -1,4 +1,4 @@
-#include <context.h>
+#include <guest.h>
 #include <gic.h>
 #include <gic_regs.h>
 #include <vdev.h>
@@ -11,7 +11,6 @@ static int32_t vdev_hvc_ping_write(struct arch_vdev_trigger_info *info,
                         struct arch_regs *regs)
 {
     printh("[hyp] _hyp_hvc_service:ping\n\r");
-    context_dump_regs(regs);
     return 0;
 }
