+++ hypervisor/vdev/vdev_timer.c
@@ -1,4 +1,4 @@
-#include <context.h>
+#include <guest.h>
 #include <gic.h>
 #include <gic_regs.h>
 #include <vdev.h>
@@ -37,7 +37,7 @@ static hvmm_status_t vdev_vtimer_access_handler(uint32_t write,
             write ? "write" : "read",
             offset, write ? *pvalue : (uint32_t) pvalue);
     hvmm_status_t result = HVMM_STATUS_BAD_ACCESS;
-    unsigned int vmid = context_current_vmid();
+    unsigned int vmid = guest_current_vmid();
     if (!write) {
         /* READ */
         switch (offset) {
@@ -101,7 +101,7 @@ static int32_t vdev_vtimer_check(struct arch_vdev_trigger_info *info,
 
 void callback_timer(void *pdata)
 {
-    vmid_t vmid = context_current_vmid();
+    vmid_t vmid = guest_current_vmid();
 
     if (_timer_status[vmid] == 0)
         virq_inject(vmid, VTIMER_IRQ, 0, 0);
