+++ platform-device/cortex_a15x2_arndale/k-hypervisor-config.h
@@ -7,5 +7,6 @@
 
 #define NUM_GUESTS_STATIC       2
 #define COUNT_PER_USEC (CFG_CNTFRQ/USEC)
+#define GUEST_SCHED_TICK 100000
 
 #endif  /* KHYPERVISOR_CONFIG_H */
